pyproject.toml
[project]
name = "query-fan-out"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "bs4>=0.0.2",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "opik>=1.7.40",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.46.0",
]
